# backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: NarrateDAO API
  version: 0.1.0
  description: 最小信任的去中心化投票解决方案 API
  contact:
    name: NarrateDAO Team
    url: https://github.com/NarrateDAO/narrate-dao
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: 开发环境
  - url: https://api.narratedao.com
    description: 生产环境
...
paths:
...
  /api/v1/weights/calculate:
    post:
      summary: 计算用户投票权重
      description: 根据提供的策略计算特定用户的投票权重
      tags:
        - 权重服务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - voterAddress
                - strategies
              properties:
                voterAddress:
                  type: string
                  description: 投票者的以太坊地址
                  example: "0x1234567890123456789012345678901234567890"
                strategies:
                  type: array
                  description: 权重计算策略数组
                  items:
                    type: object
                    required:
                      - type
                      - params
                    properties:
                      type:
                        type: string
                        enum: [erc20-balance-of, erc721-balance-of, fixed-weight]
                        description: 策略类型
                      params:
                        type: object
                        description: 策略参数
                blockNumber:
                  type: integer
                  description: 区块高度（可选，默认为最新区块）
                  example: 15000000
      responses:
        200:
          description: 成功计算权重
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: 请求验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/weights/strategies:
    get:
      summary: 获取可用的权重计算策略
      description: 返回系统支持的所有权重计算策略类型
      tags:
        - 权重服务
      responses:
        200:
          description: 成功返回策略列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        500:
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/weights/erc20/{tokenAddress}/{voterAddress}:
    get:
      summary: 获取ERC20代币余额
      description: 查询特定用户的ERC20代币余额
      tags:
        - 权重服务
      parameters:
        - in: path
          name: tokenAddress
          required: true
          schema:
            type: string
          description: ERC20代币合约地址
        - in: path
          name: voterAddress
          required: true
          schema:
            type: string
          description: 用户地址
        - in: query
          name: blockNumber
          schema:
            type: integer
          description: 区块高度（可选，默认为最新区块）
      responses:
        200:
          description: 成功返回代币余额
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: 请求验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/weights/erc721/{tokenAddress}/{voterAddress}:
    get:
      summary: 获取ERC721代币余额
      description: 查询特定用户的ERC721代币（NFT）持有数量
      tags:
        - 权重服务
      parameters:
        - in: path
          name: tokenAddress
          required: true
          schema:
            type: string
          description: ERC721代币合约地址
        - in: path
          name: voterAddress
          required: true
          schema:
            type: string
          description: 用户地址
        - in: query
          name: blockNumber
          schema:
            type: integer
          description: 区块高度（可选，默认为最新区块）
      responses:
        200:
          description: 成功返回NFT持有数量
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: 请求验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
  # 审计服务接口
...
components:
  securitySchemes:
    web3Signature:
      type: apiKey
      in: header
      name: signature
      description: 使用Web3钱包生成的签名
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          description: 状态标识
          example: error
        message:
          type: string
          description: 错误消息
        errors:
          type: array
          description: 详细错误信息
          items:
            type: object
            properties:
              field:
                type: string
                description: 错误字段
              message:
                type: string
                description: 错误描述
    
    # 成功响应
    Success:
      type: object
      properties:
        status:
          type: string
          description: 状态标识
          example: success
        message:
          type: string
          description: 成功消息
        data:
          type: object
          description: 响应数据 